feat: files responsive design
