+++ src/components/Cards/DriveCard/DriveCardIcon/DriveCardIcon.tsx
@@ -11,16 +11,18 @@ interface DriveCardIconProps {
 const DriveCardIcon: FC<DriveCardIconProps> = ({ type, fileExtention }) => {
   return (
     <div className="">
-      <span className="text-color-accents-grey-lavendar dark:text-color-shade-dark-1-night z-10">
-        {type === 'folder' ? (
-          <FolderIcon className="inline-block mb-8" />
-        ) : (
-          <FileIcon />
-        )}
+      <span className="flex text-color-accents-grey-lavendar dark:text-color-shade-dark-1-night z-10">
+        <div className="mx-auto lg:mx-0">
+          {type === 'folder' ? (
+            <FolderIcon className="inline-block mb-8" />
+          ) : (
+            <FileIcon />
+          )}
+        </div>
       </span>
 
       {type === 'file' ? (
-        <span className="block -mt-10 mb-10 ml-3 text-color-shade-black-day dark:text-color-accents-grey-pastel uppercase z-20">
+        <span className="block -mt-10 mb-10 ml-0 lg:ml-3 text-center lg:text-center text-color-shade-black-day dark:text-color-accents-grey-pastel uppercase z-20">
           {fileExtention}
         </span>
       ) : null}
