+++ src/components/Modals/SideModal/SideModal.tsx
@@ -13,6 +13,7 @@ interface SideModalProps {
     dark: ReactNode;
   };
   headerTitle: string;
+  className?: string;
   children: ReactChild | ReactChild[];
 }
 
@@ -21,6 +22,7 @@ const SideModal: FC<SideModalProps> = ({
   closeModal,
   headerIcon,
   headerTitle,
+  className,
   children,
 }) => {
   const { theme } = useContext(ThemeContext);
@@ -30,7 +32,9 @@ const SideModal: FC<SideModalProps> = ({
       {showModal ? (
         <div className="fixed inset-0 z-30" onClick={closeModal}>
           <div
-            className="absolute w-98 h-screen top-0 right-0 py-10 px-10 bg-color-shade-dark-1-day dark:bg-color-shade-dark-3-night z-50 overflow-scroll no-scroll-bar"
+            className={`absolute w-98 h-screen top-0 right-0 py-10 px-10 bg-color-shade-dark-1-day dark:bg-color-shade-dark-3-night z-50 overflow-scroll no-scroll-bar${
+              className ? ` ${className}` : ''
+            }`}
             onClick={(e) => e.stopPropagation()}
           >
             <div className="flex justify-between items-center">
